refactor: apply clean code and REST patterns to controllers and services

+++ src/main/java/com/movieflix/controller/CategoryController.java
@@ -1,14 +1,17 @@
 package com.movieflix.controller;
 
 
+import com.movieflix.dto.request.CategoryRequestDTO;
 import com.movieflix.dto.response.CategoryResponseDTO;
 import com.movieflix.entity.Category;
+import com.movieflix.mapper.CategoryMapper;
 import com.movieflix.service.CategoryService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController()
 @RequestMapping("/movieflix/category")
@@ -21,26 +24,33 @@ public class CategoryController {
     }
 
     @PostMapping
-    public ResponseEntity<CategoryResponseDTO> saveCategory(@RequestBody Category category){
-        CategoryResponseDTO categoryResponseDTO = categoryService.saveCategory(category);
-        return ResponseEntity.status(HttpStatus.CREATED).body(categoryResponseDTO);
+    public ResponseEntity<CategoryResponseDTO> save(@RequestBody CategoryRequestDTO categoryRequestDTO){
+        Category savedCategory = categoryService.save(CategoryMapper.toEntity(categoryRequestDTO));
+        return ResponseEntity.status(HttpStatus.CREATED).body(CategoryMapper.toResponse(savedCategory));
     }
 
     @GetMapping
-    public ResponseEntity<List<CategoryResponseDTO>> getAllCategories(){
-        List<CategoryResponseDTO> all = categoryService.findAll();
-        return ResponseEntity.status(HttpStatus.OK).body(all);
+    public ResponseEntity<List<CategoryResponseDTO>> findAll(){
+        return ResponseEntity.ok(categoryService.findAll()
+                .stream().map(CategoryMapper::toResponse)
+                .toList());
+
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CategoryResponseDTO> getCategoryById(@PathVariable Long id){
-        CategoryResponseDTO categoryResponseDTO = categoryService.findById(id);
-        return ResponseEntity.status(HttpStatus.OK).body(categoryResponseDTO);
+    public ResponseEntity<CategoryResponseDTO> findById(@PathVariable Long id){
+        return categoryService.findById(id).map(category ->
+                ResponseEntity.ok(CategoryMapper.toResponse(category)))
+                .orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCategoryById(@PathVariable Long id){
-        categoryService.deleteById(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    public ResponseEntity<Void> deleteById(@PathVariable Long id){
+        Optional<Category> categorySearched = categoryService.findById(id);
+        if(categorySearched.isPresent()){
+            categoryService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
