+++ src/main/java/com/movieflix/controller/StreamingController.java
@@ -2,9 +2,9 @@ package com.movieflix.controller;
 
 
 import com.movieflix.dto.request.StreamingRequestDTO;
-import com.movieflix.dto.response.CategoryResponseDTO;
 import com.movieflix.dto.response.StreamingResponseDTO;
-import com.movieflix.entity.Category;
+import com.movieflix.entity.Streaming;
+import com.movieflix.mapper.StreamingMapper;
 import com.movieflix.service.StreamingService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -23,26 +23,33 @@ public class StreamingController {
     }
 
     @PostMapping
-    public ResponseEntity<StreamingResponseDTO> saveCategory(@RequestBody StreamingRequestDTO streaming){
-        StreamingResponseDTO categoryResponseDTO = streamingService.saveCategory(streaming);
-        return ResponseEntity.status(HttpStatus.CREATED).body(categoryResponseDTO);
+    public ResponseEntity<StreamingResponseDTO> save(@RequestBody StreamingRequestDTO streaming){
+        Streaming streamingSaved = streamingService.save(StreamingMapper.toEntity(streaming));
+        return ResponseEntity.status(HttpStatus.CREATED).body(StreamingMapper.toResponse(streamingSaved));
     }
 
     @GetMapping
-    public ResponseEntity<List<StreamingResponseDTO>> getAllCategories(){
-        List<StreamingResponseDTO> all = streamingService.findAll();
-        return ResponseEntity.status(HttpStatus.OK).body(all);
+    public ResponseEntity<List<StreamingResponseDTO>> findAll(){
+        return ResponseEntity.ok(streamingService.findAll()
+                .stream()
+                .map(StreamingMapper::toResponse).toList());
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<StreamingResponseDTO> getCategoryById(@PathVariable Long id){
-        StreamingResponseDTO streamingResponseDTO = streamingService.findById(id);
-        return ResponseEntity.status(HttpStatus.OK).body(streamingResponseDTO);
+    public ResponseEntity<StreamingResponseDTO> findById(@PathVariable Long id){
+        return streamingService.findById(id).map(
+                streaming -> ResponseEntity.ok(StreamingMapper.toResponse(streaming)))
+                .orElse(ResponseEntity.notFound().build());
+
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCategoryById(@PathVariable Long id){
-        streamingService.deleteById(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    public ResponseEntity<Void> delete(@PathVariable Long id){
+        if(streamingService.streamingExists(id)){
+            streamingService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 }
